Time: 0 ms (100.00%) | Memory: 111.2 MB (74.50%) - LeetSync

+++ 2432-number-of-zero-filled-subarrays/number-of-zero-filled-subarrays.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
         long long cnt =0;
-        long long streak=0;
+        long streak=0;
         for(int num:nums){
             if(num==0){
                 streak++;
